Makefile: out-of-source build for GNU Make.
Build babeld in a designated build directory:

    make BUILD=build/

Beware! The trailing slash is mandatory.

Pros of out-of-source tree builds:
- object and dependency files are constrained in one location;
- it's easier to have multiple builds of babeld, for instance when
  compiling for multiple architectures, or switching between debug and
  release modes.

An usual call to `make` will build babeld in the current directory. No
breaking change on the current developer and packager workflow.

BSD Make does not support "pattern rules", and so will not support
out-of-source builds. Use it as usual.

Tested with GNU Make and BSD Make.